+++ packages/plugins/Draw/src/store/createInteractions/createModifyInteractions.ts
@@ -4,7 +4,7 @@ import { PolarActionContext } from '@polar/lib-custom-types'
 import { Collection, Feature, Map } from 'ol'
 import { CreateInteractionsPayload, DrawGetters, DrawState } from '../../types'
 import { makeLocalSelector } from './localSelector'
-import { getSchnaps } from './getSnaps'
+import { getSnaps } from './getSnaps'
 
 const createModify = (
   map: Map,
@@ -76,7 +76,7 @@ export default function (
   })
   return [
     createModify(rootGetters.map, drawLayer),
-    ...getSchnaps(
+    ...getSnaps(
       rootGetters.map,
       rootGetters.configuration?.draw?.snapTo || []
     ),