+++ packages/plugins/Draw/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Feature: Add a lasso mode that allows copying up features from a vector layer that are contained within the user's hand drawn polygon. This also adds the fields `addLoading`, `removeLoading`, and `toastAction` for usage in the `lassos`.
 - Feature: Interactions requiring dragging are now marked with the POLAR flag `_isPolarDragLikeInteraction`.
 - Feature: An action `setInteractions` has been added to allow clients to bring their own geometry operations.
+- Feature: Expose the `getSnaps` function. Intended to be used together with the `setInteractions` action only.
 
 ## 3.0.0